rename: Image 관련 클래스 디렉토리 변경
