+++ be/src/main/java/kr/codesquad/secondhand/api/image/util/S3ImageUploader.java
@@ -1,11 +1,11 @@
-package kr.codesquad.secondhand.api.product.service;
+package kr.codesquad.secondhand.api.image.util;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import java.io.IOException;
 import java.net.URL;
 import java.util.UUID;
-import kr.codesquad.secondhand.api.product.exception.ImageUploadFailedException;
+import kr.codesquad.secondhand.api.image.excpetion.ImageUploadFailedException;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
