+++ be/src/main/java/kr/codesquad/secondhand/api/product/service/ProductFacadeService.java
@@ -6,6 +6,7 @@ import java.util.Map;
 import kr.codesquad.secondhand.api.address.domain.Address;
 import kr.codesquad.secondhand.api.address.service.AddressService;
 import kr.codesquad.secondhand.api.category.domain.Category;
+import kr.codesquad.secondhand.api.image.service.ImageService;
 import kr.codesquad.secondhand.api.member.domain.Member;
 import kr.codesquad.secondhand.api.member.service.MemberService;
 import kr.codesquad.secondhand.api.product.domain.Product;
